more cleanup and testing scripts
