+++ weibull.py
@@ -1,10 +1,8 @@
 import numpy as np
 import torch
-from weibull.backend_pytorch import fit
-import timeit
-import os, sys
 from pynvml import *
 
+from backend_weibull import fit
 
 def determine_splits(inputTensor, tailSize, isSorted = 0):
     
@@ -55,6 +53,7 @@ def weibullFitting(dataTensor, tailSize, sign, isSorted = 0):
   processedTensor = sortedTensor + trnaslateAmoutTensor - smallScoreTensor
   wbFits = fit(processedTensor)
   
+  print ("signTensor",wbFits.dtype,signTensor.dtype,smallScoreTensor.dtype)
   resultTensor = torch.cat([wbFits[:,1].unsqueeze(1),wbFits[:,0].unsqueeze(1),signTensor,trnaslateAmoutTensor,smallScoreTensor], dim=1)
 
   return resultTensor
@@ -62,47 +61,7 @@ def weibullFitting(dataTensor, tailSize, sign, isSorted = 0):
 def FitLow(data, tailSize, isSorted = 0):
   return weibullFitting(data, tailSize, -1, isSorted)
   
-  
 def FitHigh(data, tailSize, isSorted = 0):
   return weibullFitting(data, tailSize, 1, isSorted)
   
 
-def test_weibullFit():
-  # getting the tensor ready to get the weibull fits for all instances 
-  fileName = 'sample_data/weibulls_example_protocol2.npy'
-  data = np.load(fileName, allow_pickle=True)
-  
-  numInstances = 10
-  dataSize = 40000
-  tailSize = 2500
-  distance_multiplier = 0.5
-  
-  data1 = np.zeros(shape=(numInstances,dataSize), dtype=np.float64)
-  weibull_fits_libmr = np.zeros(shape=(numInstances,5), dtype=np.float64)
-  
-  for k in range(numInstances):
-    data1[k,:] = distance_multiplier * data[k][0]
-    weibull_fits_libmr[k,:] = data[k][1]
-  
-  dataTensor = torch.from_numpy(data1)
-  dataTensor = dataTensor.cuda()
-  
-  # calling the weibull fit for the tensor
-  result = FitLow(dataTensor, tailSize, 0)
-  
-  print(result)
-  
-  return
-  
-def test_determine_splits():
-  dataTensor = torch.rand(size=(1000,40000), dtype=torch.float64)
-  dataTensor = dataTensor.cuda()
-  splits = determine_splits(dataTensor, 2500, 0)
-  print(splits)
-  
-  
-if __name__ == '__main__':
-  test_weibullFit()
-  #test_determine_splits()
-  
-
